pre-css

+++ app/views/coins/coins.php
@@ -31,6 +31,7 @@ if (isset($data['addedCoinName'])) : ?>
     <tr>
         <th>Coin name</th>
         <th>Coin symbol</th>
+        <th>API ID</th>
         <th>Latest price</th>
         <th>Edit</th>
         <th>Delete</th>
@@ -42,19 +43,22 @@ if (isset($data['addedCoinName'])) : ?>
         <tr>
             <td><?= $row->name ?></td>
             <td><?= $row->symbol ?></td>
+            <td><?= $row->api_id ?></td>
             <td><?= $row->latestprice ?></td>
             <td><a href="http://<?php echo URLROOT; ?>/coins/edit/<?= $row->id ?>">Edit</a></td>
-            <td><a href="http://<?php echo URLROOT; ?>/coins/delete/<?= $row->id ?>">Delete</a></td>
+            <td><a href="http://<?php echo URLROOT; ?>/coins/delete/<?= $row->id ?>" class="confirmDelete">Delete</a></td>
         </tr>
 
     <?php endforeach; ?>
 </table>
 
-<form action="http://<?php echo URLROOT?>/coins/add" method="GET">
+<form action="http://<?php echo URLROOT?>/coins/add" method="POST">
 <label for="coinname">Coin: </label>
 <input type="text" name="coinname" id="coinname" placeholder="Coin name" required>
 <label for="coinsymbol">Symbol: </label>
 <input type="text" name="coinsymbol" id="coinsymbol" placeholder="Coin symbol" required>
+<label for="apiID">Symbol: </label>
+<input type="text" name="apiID" id="apiID" placeholder="API ID" required>
 <input type="submit" value="Add">
 
 </form>
